requirements.txt
flask
pandas
supabase
python-dotenv
